requirements.txt
fastapi
uvicorn
pandas
plotly
gradio
gradio_modal
uvicorn
httpx
pydantic
supabase
python-dotenv
authentication
bcrypt